backend/tests/requirements.txt
python-dotenv
requests
sqlalchemy
pytest-ordering
pytest
pytest-asyncio
httpx
pydantic-settings
asyncpg
psycopg2-binary
